Split codegen Tests
